backport numpy 2.0 compatibilty for a maintenance release 1.4.2

+++ obspy/io/mseed/util.py
@@ -1609,7 +1609,7 @@ def shift_time_of_file(input_file, output_file, timeshift):
         # This should rarely be the case.
         if current_time_shift == 0 and is_time_correction_applied:
             # This sets bit 2 of the activity flags to 0.
-            current_record[36] = current_record[36] & (~2)
+            current_record[36] = np.int64(current_record[36]) & (~2)
             is_time_correction_applied = False
         # This is the case if the time correction has been applied. This
         # requires some more work by changing both, the actual time and the
